cad-pulse: improve profile selection
Starting with v1.2.0, PipeWire exposes one instance of each profile per
possible port combination, with the priority of each instance being
calculated from the priority of each included port. Each instance is
also marked as available only if all included ports are available.

This causes `callaudiod` to fail detecting the profile to be used as
profile names not only include the profile itself but also the list of
included ports.

Therefore, we should not use strict equality for profile names but only
check the beginning of the string. In order to select the correct
instance we switch to selecting the highest priority available profile,
the same way things are done for ports selection.

Notes:
* this breaks speaker selection as we'll generally switch to an
  earpiece-only profile instance
* devices without a `VoiceCall` profile (e.g. Librem 5) might break as
  well, would be best to create such a profile for those

Fixes #35